refactor: mainAccount module 추출
